+++ src/main/scala/com/wenjunhuang/codeepiphany/atcoder/services/AtCoderSubmissionService.scala
@@ -31,10 +31,12 @@ import com.wenjunhuang.codeepiphany.services.{
   ChallengeRepository,
   WebViewStyleProvider
 }
+import com.wenjunhuang.codeepiphany.services.console.MessageSeg
 import com.wenjunhuang.codeepiphany.services.http.HttpClientManager
 import com.wenjunhuang.codeepiphany.settings.ChallengeSettings.ChallengeSettingsStateItem
 import com.wenjunhuang.codeepiphany.utils.jcef.BaseJCefWebView
 import com.wenjunhuang.codeepiphany.utils.syntax.*
+import com.wenjunhuang.codeepiphany.vfs.WebPreviewVirtualFile
 
 class AtCoderSubmissionService(project: Project) extends BaseSubmissionService(project, AtCoder) {
   override type SubmissionRequest  = Request
@@ -58,7 +60,7 @@ class AtCoderSubmissionService(project: Project) extends BaseSubmissionService(p
       basicInfo.problemId,
       basicInfo.languageId,
       processedCode,
-      showCloudflare(basicInfo)
+      getCSRFAndCloudflare(basicInfo)
     )
 
   override protected def reportSubmitResult(
@@ -70,9 +72,30 @@ class AtCoderSubmissionService(project: Project) extends BaseSubmissionService(p
   ): IO[Unit] = {
     lastResponseInfo.result match
       case SubmissionResult.Success =>
-        console.info(project, AtCoder, PluginBundle.message("submission.passed"))
+        console.info(
+          project,
+          CodeDojo.AtCoder,
+          MessageSeg.Hyperlink(
+            PluginBundle.message("submissionResult.viewDetails"),
+            { link =>
+              AtCoderSubmissionService.showSubmissionDetails(project, lastResponse.submissionId, basicInfo.contestId)
+            }
+          ),
+          "\n",
+          PluginBundle.message("submission.passed") + "\n${lastResponse.message}"
+        )
       case _ =>
-        console.error(project, AtCoder, s"${lastResponseInfo.result.show}\n${lastResponseInfo.message}")
+        console.error(
+          project,
+          MessageSeg.Hyperlink(
+            PluginBundle.message("submissionResult.viewDetails"),
+            { link =>
+              AtCoderSubmissionService.showSubmissionDetails(project, lastResponse.submissionId, basicInfo.contestId)
+            }
+          ),
+          "\n",
+          s"${lastResponseInfo.result.show}\n${lastResponse.message}"
+        )
   }
 
   private def createRequest(item: ChallengeSettingsStateItem, client: DSLContext): Request = {
@@ -107,20 +130,24 @@ class AtCoderSubmissionService(project: Project) extends BaseSubmissionService(p
   private def resolveLanguageId(language: Language, version: LanguageVersion): Option[String] =
     AtCoderSettingsConfigurable.ATCODER_LANGUAGES.get((language, version))
 
-  private def showCloudflare(request: Request): IO[String] = {
+  private def getCSRFAndCloudflare(request: Request): IO[(String, String)] = {
     HttpClientManager
       .getCookiesForHost(CodeDojo.AtCoder.domain)
       .flatMap { cookies =>
-        IO.async_[String] { cb =>
+        IO.async_[(String, String)] { cb =>
           val dialog           = DialogBuilder(myProject)
           val browserComponent = BaseJCefWebView.createBrowser()
 
           val jsQuery = JBCefJSQuery.create(browserComponent.asInstanceOf[JBCefBrowserBase])
           jsQuery.addHandler { (s: String) =>
+            val (csrf, turnstile) = s.split(",") match {
+              case Array(c, t) => (c, t)
+              case _           => throw new IllegalArgumentException("Invalid CSRF and Turnstile response format")
+            }
             ApplicationManager.getApplication.invokeLater(
               { () =>
                 dialog.getDialogWrapper.close(0, true)
-                cb(Right(s))
+                cb(Right((csrf, turnstile)))
               },
               ModalityState.any()
             )
@@ -160,10 +187,13 @@ class AtCoderSubmissionService(project: Project) extends BaseSubmissionService(p
                          |   mask.style = "position: fixed; top: 0; left: 0; width: 100vw; height: 100vh; background-color: ${WebViewStyleProvider.DEFAULT.panelBackground
                           .webRgba(1.0)}; z-index: 9999;display: flex; align-items: center; justify-content: center;";
                          |   document.body.append(mask);
+                         |   const csrf = document.querySelector("input[name='csrf_token']").value;
                          |   turnstile.render('#cf-turnstile-mask', {
                          |      sitekey: '0x4AAAAAAA6HJUmmLP7mLxx0',
-                         |      callback: function(result) {
-                         |        console.log('Success:',result);
+                         |      theme: ${if (WebViewStyleProvider.DEFAULT.isDarkMode) "'dark'" else "'light'"},
+                         |      callback: function(turnstile) {
+                         |        const result = csrf + "," + turnstile;
+                         |        console.log("CSRF and Turnstile result: " , result);
                          |        ${jsQuery.inject("result")}
                          |   }});
                          |}
@@ -219,4 +249,22 @@ class AtCoderSubmissionService(project: Project) extends BaseSubmissionService(p
     langVer: LanguageVersion,
     languageId: String
   )
+}
+object AtCoderSubmissionService {
+  def showSubmissionDetails(project: Project, submissionId: String, contestId: String): Unit = {
+    HttpClientManager
+      .getCookiesForHost(CodeDojo.AtCoder.domain)
+      .flatMap { cookies =>
+        IO.delay {
+          val file = new WebPreviewVirtualFile(
+            s"https://${CodeDojo.AtCoder.domain.toString}/contests/$contestId/submissions/$submissionId/",
+            CodeDojo.AtCoder.domain.toString,
+            cookies,
+            submissionId
+          )
+          WebPreviewVirtualFile.openEditor(file, project)
+        }.evalOnEDTAny()
+      }
+      .unsafeRunAndForget()
+  }
 }